move files out of bhswz module
