--- src/bhswz/mod.rs
@@ -1,10 +1,0 @@
-mod swz_random;
-use swz_random::SwzRandom;
-pub mod swz_reader;
-pub use swz_reader::SwzReader;
-pub mod swz_writer;
-pub use swz_writer::SwzWriter;
-pub mod swz_filename;
-pub use swz_filename::get_swz_file_name;
-mod swz_utils;
-use swz_utils::*;
