+++ src/swz_reader.rs
