+++ src/swz_writer.rs
